Quality of Life Improvements for the Gui Editor
This makes a number of improvements to the Gui Editor: Reduced the number of fields saved for a taml gui file. Added the "Is Container" field back to the inspector. Added support for Hidden and Locked controls. Fixed the text edit boxes to apply their value on enter and revert values on escape. Fixes the editor fields for fluid colors used by the sprite control.